requirements.txt
joblib
torchvision
torch
pandas
xgboost
matplotlib
scikit-learn
opencv-python
numpy
pillow
